fixsessionsmigrations

+++ db/migrate/20130416024901_change_score_value_to_double.rb
@@ -1,6 +1,6 @@
 class ChangeScoreValueToDouble < ActiveRecord::Migration
   def up
-  	change_column :scores, :score_value, :double
+  	change_column :scores, :score_value, :float
   end
 
   def down
